requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
djangorestframework==3.12.4
idna==3.3
iniconfig==1.1.1
packaging==21.3
Pillow==8.4.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.3
requests==2.26.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
